Fix create_project.tcl BD detection
The create_project.tcl script was using get_bd_designs to find the name
of the block diagram project. This will break if an IP core used in the
design contains a sub-block diagram, as is the case with the AXI
smartconnect IP.

The script now identifies the name of the block diagram by identifying
the name of the folder found in src/bd, which will match the BD name.